Resolution changes, but paddles don't. Toggle still doesn't work.
I am tired.

Will need to come back to it and figure out why the simple toggle fullscreen isn't working. Reverted the toggle back to what it first was.
The paddles don't change with the resolution size either, don't even have the energy to figure that out.
But at least the resolution sizes match your monitor's refresh rate and filters it, so it doesn't give you a billion options to choose from, AND is organized and not a jumbled mess.
sigh~